Move solution text to frontend for easy deployment of changes
