+++ src/Diagnostics.ModelsAndUtils/Utilities/SolutionText.cs
@@ -1,0 +1,13 @@
+﻿namespace Diagnostics.ModelsAndUtils.Utilities
+{
+    /// <summary>
+    /// Signify text values for <see cref="Solution"/>s that are added in the Solution's rendering process.
+    /// </summary>
+    public enum SolutionText
+    {
+        AppRestartDescription,
+        RestartInstructions,
+        UpdateSettingsInstructions,
+        UpdateSettingsDescription
+    }
+}
